tools: Fix warnings.

+++ tools/divine.cpp
@@ -158,7 +158,7 @@ struct Main {
 
         {
             std::ostringstream execCommStr;
-            for(size_t i = 0; i < argc; ++i)
+            for( int i = 0; i < argc; ++i )
                 execCommStr << argv[i] << " ";
             report.execCommand = execCommStr.str();
         }
